+++ tests/test_visitors/test_ast/test_general/test_wrong_import/test_future_imports.py
@@ -3,7 +3,7 @@
 import pytest
 
 from wemake_python_styleguide.constants import FUTURE_IMPORTS_WHITELIST
-from wemake_python_styleguide.visitors.ast.wrong_import import (
+from wemake_python_styleguide.visitors.ast.general.wrong_import import (
     FutureImportViolation,
     WrongImportVisitor,
 )