+++ tests/test_visitors/test_ast/test_general/test_wrong_import/test_relative_imports.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from wemake_python_styleguide.visitors.ast.wrong_import import (
+from wemake_python_styleguide.visitors.ast.general.wrong_import import (
     LocalFolderImportViolation,
     WrongImportVisitor,
 )