+++ tests/test_visitors/test_ast/test_general/test_wrong_keyword/test_del.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from wemake_python_styleguide.visitors.ast.wrong_keyword import (
+from wemake_python_styleguide.visitors.ast.general.wrong_keyword import (
     WrongKeywordViolation,
     WrongKeywordVisitor,
 )