+++ tests/test_visitors/test_ast/test_general/test_wrong_name/test_function_names.py
@@ -4,7 +4,7 @@ import string
 
 import pytest
 
-from wemake_python_styleguide.visitors.ast.wrong_name import (
+from wemake_python_styleguide.visitors.ast.general.wrong_name import (
     BAD_VARIABLE_NAMES,
     PrivateNameViolation,
     TooShortVariableNameViolation,
@@ -12,27 +12,25 @@ from wemake_python_styleguide.visitors.ast.wrong_name import (
     WrongVariableNameViolation,
 )
 
-static_attribute = """
-class Test:
-    {0} = None
+function_bad_name = """
+def {0}(): ...
 """
 
-instance_attribute = """
-class Test:
-    def __init__(self):
-        self.{0} = 123
+method_bad_name = """
+class Input:
+    def {0}(self): ...
 """
 
 
 @pytest.mark.parametrize('bad_name', BAD_VARIABLE_NAMES)
 @pytest.mark.parametrize('code', [
-    static_attribute,
-    instance_attribute,
+    function_bad_name,
+    method_bad_name,
 ])
-def test_wrong_attributes_names(
+def test_wrong_function_names(
     assert_errors, parse_ast_tree, bad_name, code, default_options,
 ):
-    """Testing that attribute can not have blacklisted names."""
+    """Testing that function can not have blacklisted names."""
     tree = parse_ast_tree(code.format(bad_name))
 
     visitor = WrongNameVisitor(default_options, tree=tree)
@@ -43,13 +41,13 @@ def test_wrong_attributes_names(
 
 @pytest.mark.parametrize('short_name', string.ascii_letters)
 @pytest.mark.parametrize('code', [
-    static_attribute,
-    instance_attribute,
+    function_bad_name,
+    method_bad_name,
 ])
-def test_too_short_attribute_names(
+def test_too_short_function_names(
     assert_errors, parse_ast_tree, short_name, code, default_options,
 ):
-    """Testing that attribute can not have too short names."""
+    """Testing that function can not have too short names."""
     tree = parse_ast_tree(code.format(short_name))
 
     visitor = WrongNameVisitor(default_options, tree=tree)
@@ -59,14 +57,14 @@ def test_too_short_attribute_names(
 
 
 @pytest.mark.parametrize('code', [
-    static_attribute,
-    instance_attribute,
+    function_bad_name,
+    method_bad_name,
 ])
-def test_private_attribute_names(
+def test_private_function_names(
     assert_errors, parse_ast_tree, code, default_options,
 ):
-    """Testing that attribute can not have private names."""
-    tree = parse_ast_tree(code.format('__private_name'))
+    """Testing that function can not have private names."""
+    tree = parse_ast_tree(code.format('__hidden'))
 
     visitor = WrongNameVisitor(default_options, tree=tree)
     visitor.run()
@@ -75,20 +73,20 @@ def test_private_attribute_names(
 
 
 @pytest.mark.parametrize('correct_name', [
-    'correct_name',
+    'my_function',
     'xy',
-    'test1',
+    'test',
     '_protected',
     '__magic__',
 ])
 @pytest.mark.parametrize('code', [
-    static_attribute,
-    instance_attribute,
+    function_bad_name,
+    method_bad_name,
 ])
-def test_correct_attribute_name(
-    assert_errors, parse_ast_tree, code, correct_name, default_options,
+def test_correct_function_names(
+    assert_errors, parse_ast_tree, correct_name, code, default_options,
 ):
-    """Testing that attribute can have normal names."""
+    """Testing that function can have normal names."""
     tree = parse_ast_tree(code.format(correct_name))
 
     visitor = WrongNameVisitor(default_options, tree=tree)