+++ tests/test_visitors/test_ast/test_general/test_wrong_name/test_module_metadata.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from wemake_python_styleguide.visitors.ast.wrong_name import (
+from wemake_python_styleguide.visitors.ast.general.wrong_name import (
     BAD_MODULE_METADATA_VARIABLES,
     WrongModuleMetadataViolation,
     WrongModuleMetadataVisitor,