+++ wemake_python_styleguide/errors/complexity.py
@@ -9,6 +9,9 @@ https://en.wikipedia.org/wiki/The_Magical_Number_Seven,_Plus_or_Minus_Two
 That's how many objects we can keep in our memory at a time.
 We try hard not to exceed the limit.
 
+You can also find interesting reading about "Cognitive complexity":
+https://www.sonarsource.com/docs/CognitiveComplexity.pdf
+
 What we call "design flaws":
 
 1. Complex code (there are a lof of complexity checks!)
@@ -19,6 +22,7 @@ Note:
     Simple is better than complex.
     Complex is better than complicated.
     Flat is better than nested.
+    Namespaces are one honking great idea -- let's do more of those!
 
 """
 
@@ -197,7 +201,8 @@ class TooManyElifsViolation(ASTStyleViolation):
 
     """
 
-    error_template = '{0} Found too many "{1}" branches'
+    should_use_text = False
+    error_template = '{0} Found too many `elif` branches'
     code = 'Z204'
 
 
@@ -230,7 +235,7 @@ class TooManyExpressionsViolation(ASTStyleViolation):
 
     Reasoning:
         When there are too many expression it means that this specific
-        function does too many things at once. It has too many logics.
+        function does too many things at once. It has too many logic.
 
     Solution:
         Split function into several functions, refactor your API.
@@ -256,7 +261,7 @@ class TooDeepNestingViolation(ASTStyleViolation):
         suited to handle such construction.
 
     Solution:
-        We need to refactor our complex construction into simplier ones.
+        We need to refactor our complex construction into simpler ones.
         We can use new functions or different constructions.
 
     This rule is configurable with ``--max-offset-blocks``.
@@ -270,7 +275,7 @@ class TooDeepNestingViolation(ASTStyleViolation):
     code = 'Z207'
 
 
-class TooManyModuleMembersViolation(ASTStyleViolation):
+class TooManyModuleMembersViolation(SimpleStyleViolation):
     """
     This rule forbids to have many classes and functions in a single module.
 
@@ -293,11 +298,12 @@ class TooManyModuleMembersViolation(ASTStyleViolation):
 
     """
 
-    error_template = '{0} Found too many members "{1}"'
+    should_use_text = False
+    error_template = '{0} Found too many members'
     code = 'Z208'
 
 
-class TooManyMethodsViolation(ASTStyleViolation):
+class TooManyMethodsViolation(SimpleStyleViolation):
     """
     This rule forbids to have many methods in a single class.
 
@@ -336,7 +342,7 @@ class LineComplexityViolation(ASTStyleViolation):
     We are using Jones Complexity algorithm to count complexity.
     What is Jones Complexity? It is a simple yet power method to count
     the number of ``ast`` nodes per line.
-    If the complexity of a single line is higher than a tresshold,
+    If the complexity of a single line is higher than a threshold,
     then an error is raised.
 
     What nodes do we count? All except the following:
@@ -375,13 +381,14 @@ class JonesScoreViolation(SimpleStyleViolation):
     """
     This rule forbids to have modules with complex lines.
 
-    We are using Jones Complexity algorithm to count module score.
+    We are using Jones Complexity algorithm to count module's score.
     See
     :py:class:`~.LineComplexityViolation` for details of per-line-complexity.
-    How it is done: we count complexity per line
+    How it is done: we count complexity per line, then measuring the median
+    complexity across the lines in the whole module.
 
     Reasoning:
-        Having complex modules will decrease your code maintability.
+        Having complex modules will decrease your code maintainability.
 
     Solution:
         Refactor the module contents.
@@ -399,3 +406,40 @@ class JonesScoreViolation(SimpleStyleViolation):
     should_use_text = False
     error_template = '{0} Found module with high Jones score'
     code = 'Z211'
+
+
+class TooManyImportsViolation(SimpleStyleViolation):
+    """
+    This rule forbids to have modules with too many imports.
+
+    Namespaces are one honking great idea -- let's do more of those!
+
+    Reasoning:
+        Having too many imports without prefixes is quite expensive.
+        You have to memorize all the source locations of the imports.
+        And sometimes it is hard to remember what kind of functions and classes
+        are already injected into your context.
+
+        It is also a questionable design if a single module has a lot of
+        imports. Why a single module has so many dependencies?
+        So, it becomes too coupled.
+
+    Solution:
+        Refactor the imports to import a common namespace. Something like
+        ``from package import module`` and then
+        use it like ``module.function()``.
+
+        Or refactor your code and split the complex module into several ones.
+
+    We do not make any differences between
+    ``import`` and ``from ... import ...``.
+
+    This rule is configurable with ``--max-imports``.
+
+    Note:
+        Returns Z212 as error code
+
+    """
+
+    error_template = '{0} Found module with too many imports: {1}'
+    code = 'Z212'